requirements.txt
Django==5.0.2
djangorestframework==3.14.0
python-dotenv==1.0.1
requests==2.31.0
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
django-redis==5.4.0
Pillow==10.2.0
python-dateutil==2.8.2
django-filter==23.5
markdown==3.5.2
dj-database-url==2.1.0